requirements.txt
foolbox==3.3.3
numpy==1.24.2
pandas==1.5.3
Pillow==9.5.0
torch==2.0.0+cu118
torchvision==0.15.1+cu118
tqdm==4.65.0
